Narrow more types

+++ PKHeX.Core/Legality/Encounters/Generator/EncounterFinder.cs
@@ -161,7 +161,7 @@ public static class EncounterFinder
 
     private static string GetHintWhyNotFound(PKM pk, int gen)
     {
-        if (WasGiftEgg(pk, gen, pk.Egg_Location))
+        if (WasGiftEgg(pk, gen, (ushort)pk.Egg_Location))
             return LEncGift;
         if (WasEventEgg(pk, gen))
             return LEncGiftEggEvent;
@@ -170,9 +170,9 @@ public static class EncounterFinder
         return LEncInvalid;
     }
 
-    private static bool WasGiftEgg(PKM pk, int gen, int loc) => !pk.FatefulEncounter && gen switch
+    private static bool WasGiftEgg(PKM pk, int gen, ushort loc) => !pk.FatefulEncounter && gen switch
     {
-        3 => pk.IsEgg && pk.Met_Location == 253, // Gift Egg, indistinguishable from normal eggs after hatch
+        3 => pk.IsEgg && (byte)pk.Met_Location == 253, // Gift Egg, indistinguishable from normal eggs after hatch
         4 => Legal.GiftEggLocation4.Contains(loc) || (pk.Format != 4 && (loc == Locations.Faraway4 && pk.HGSS)),
         5 => loc is Locations.Breeder5,
         _ => loc is Locations.Breeder6,
