+++ PKHeX.Core/Legality/Restrictions/GBRestrictions.cs
@@ -12,7 +12,7 @@ namespace PKHeX.Core;
 /// </summary>
 internal static class GBRestrictions
 {
-    private static readonly HashSet<int> Stadium_GiftSpecies = new()
+    private static readonly HashSet<byte> Stadium_GiftSpecies = new()
     {
         (int)Bulbasaur,
         (int)Charmander,
@@ -34,7 +34,7 @@ internal static class GBRestrictions
     /// <summary>
     /// Species that have a catch rate value that is different from their pre-evolutions, and cannot be obtained directly.
     /// </summary>
-    internal static readonly HashSet<int> Species_NotAvailable_CatchRate = new()
+    internal static readonly HashSet<byte> Species_NotAvailable_CatchRate = new()
     {
         (int)Butterfree,
         (int)Pidgeot,
@@ -55,7 +55,7 @@ internal static class GBRestrictions
         (int)Dragonite,
     };
 
-    internal static readonly HashSet<int> Trade_Evolution1 = new()
+    internal static readonly HashSet<byte> Trade_Evolution1 = new()
     {
         (int)Kadabra,
         (int)Machoke,
@@ -83,7 +83,7 @@ internal static class GBRestrictions
         foreach (var entry in chain)
         {
             var s = entry.Species;
-            if (Species_NotAvailable_CatchRate.Contains(s))
+            if (Species_NotAvailable_CatchRate.Contains((byte)s))
                 continue;
             if (catch_rate == PersonalTable.RB[s].CatchRate || catch_rate == PersonalTable.Y[s].CatchRate)
                 return true;
@@ -94,7 +94,7 @@ internal static class GBRestrictions
         if (catch_rate == 204 && (species is (int)Krabby or (int)Kingler))
             return true;
 
-        if (catch_rate is (167 or 168) && Stadium_GiftSpecies.Contains(species))
+        if (catch_rate is (167 or 168) && Stadium_GiftSpecies.Contains((byte)species))
             return true;
 
         return false;
